+++ PyCo/System/Systems.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2018-2019 Antoine Sanner
+# Copyright 2018, 2020 Antoine Sanner
 #           2016, 2019 Lars Pastewka
 #           2016 Till Junge
 #