+++ tests/Plastic_tests.py
@@ -1,38 +1,62 @@
+#
+# Copyright 2019-2020 Lars Pastewka
+#           2018, 2020 Antoine Sanner
+# 
+# ### MIT license
+# 
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+# 
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+# 
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
 import numpy as np
-import pytest
+import os
 from scipy.optimize import bisect
+
+from NuMPI.Tools.Reduction import Reduction
+from NuMPI import MPI
+
 from PyCo.ContactMechanics import HardWall
 from PyCo.SolidMechanics import PeriodicFFTElasticHalfSpace
 from PyCo.Topography import open_topography, PlasticTopography
 from PyCo.System import make_system
-from NuMPI.Tools.Reduction import Reduction
-from PyCo.Tools.Logger import screen
 from PyCo.Topography import Topography
-from runtests.mpi import MPITest
-import os
-import PyCo
+
 
 FIXTURE_DIR = os.path.join(
     os.path.dirname(os.path.realpath(__file__)),
     'file_format_examples')
 
-def test_hard_wall_bearing_area(comm, fftengine_type):
+def test_hard_wall_bearing_area(comm):
     # Test that at very low hardness we converge to (almost) the bearing
     # area geometry
     pnp = Reduction(comm)
     fullsurface = open_topography(
         os.path.join(FIXTURE_DIR, 'surface1.out')).topography()
     nb_domain_grid_pts = fullsurface.nb_grid_pts
-    substrate = PeriodicFFTElasticHalfSpace(nb_domain_grid_pts, 1.0, fft="serial"
-                            if comm.Get_size() == 1 else "mpi", communicator=comm)
+    substrate = PeriodicFFTElasticHalfSpace(nb_domain_grid_pts, 1.0, fft='mpi', communicator=comm)
     surface = Topography(fullsurface.heights(), physical_sizes=nb_domain_grid_pts,
                          decomposition='domain',
                          subdomain_locations=substrate.topography_subdomain_locations,
                          nb_subdomain_grid_pts=substrate.topography_nb_subdomain_grid_pts,
                          communicator=substrate.communicator)
 
+    plastic_surface = PlasticTopography(surface, 1e-12)
     system = make_system(substrate,
-                         HardWall(), PlasticTopography(surface, 0.0))
+                         HardWall(), plastic_surface)
     offset = -0.002
     if comm.rank == 0:
         def cb(it, p_r, d):
@@ -45,13 +69,14 @@ def test_hard_wall_bearing_area(comm, fftengine_type):
     assert result.success
     c = result.jac > 0.0
     ncontact = pnp.sum(c)
+    assert plastic_surface.plastic_area == ncontact * surface.area_per_pt
     bearing_area = bisect(lambda x: pnp.sum((surface.heights() > x)) - ncontact,
                           -0.03, 0.03)
     cba = surface.heights() > bearing_area
     # print(comm.Get_rank())
     assert pnp.sum(np.logical_not(c == cba)) < 25
 
-def test_hardwall_plastic_nonperiodic_disp_control(comm, fftengine_type):
+def test_hardwall_plastic_nonperiodic_disp_control(comm_self):
     # test just that it works without bug, not accuracy
 
     nx, ny = 128, 128
@@ -72,7 +97,7 @@ def test_hardwall_plastic_nonperiodic_disp_control(comm, fftengine_type):
                          surface=PlasticTopography(topography=topography,
                                                    hardness=hardness),
                          young=Es,
-                         communicator=comm
+                         communicator=comm_self
                          )
 
     offsets = [1e-4]
@@ -97,7 +122,7 @@ def test_hardwall_plastic_nonperiodic_disp_control(comm, fftengine_type):
         assert sol.success
 
 
-def test_hardwall_plastic_nonperiodic_load_control(comm, fftengine_type):
+def test_hardwall_plastic_nonperiodic_load_control(comm_self):
     # test just that it works without bug, not accuracy
 
     nx, ny = 128, 128
@@ -118,7 +143,7 @@ def test_hardwall_plastic_nonperiodic_load_control(comm, fftengine_type):
                          surface=PlasticTopography(topography=topography,
                                                    hardness=hardness),
                          young=Es,
-                         communicator=comm
+                         communicator=comm_self
                          )
 
     external_forces = [0.02]