+++ tests/Topography/AutocorrelationTest.py
@@ -1,5 +1,6 @@
 #
-# Copyright 2019 Lars Pastewka
+# Copyright 2019 Antoine Sanner
+#           2019 Lars Pastewka
 # 
 # ### MIT license
 #