+++ tests/Topography/IO/OPDxImportTests.py
@@ -1,6 +1,9 @@
 #
-# Copyright 2019 Antoine Sanner
-#           2019 Kai Haase
+# Copyright 2019-2020 Michael Röttger
+#           2019-2020 Antoine Sanner
+#           2019-2020 Kai Haase
+#           2019 Lars Pastewka
+#           2019 roettger@tf.uni-freiburg.de
 # 
 # ### MIT license
 # 
@@ -139,13 +142,15 @@ class OPDxSurfaceTest(unittest.TestCase):
         self.assertEqual(channel_0.info['Height_value'], 35.85522403809594)
         self.assertEqual(channel_0.info['z_scale'], 1.0)
 
+        assert 'unit' not in channel_0.info  # there is no z unit so we cannot return a common unit here
+
         # .. mandatory keys
         self.assertEqual(channel_0.name, 'Image')
         self.assertEqual(channel_0.dim, 2)
-        self.assertAlmostEqual(channel_0.physical_sizes[0], 35.85522403809594)
-        self.assertAlmostEqual(channel_0.physical_sizes[1], 47.81942809668896)
-        self.assertAlmostEqual(channel_0.nb_grid_pts[0], 960)
-        self.assertAlmostEqual(channel_0.nb_grid_pts[1], 1280)
+        self.assertAlmostEqual(channel_0.physical_sizes[1], 35.85522403809594)
+        self.assertAlmostEqual(channel_0.physical_sizes[0], 47.81942809668896)
+        self.assertAlmostEqual(channel_0.nb_grid_pts[1], 960)
+        self.assertAlmostEqual(channel_0.nb_grid_pts[0], 1280)
 
         #
         # Channel 1: Raw
@@ -157,15 +162,17 @@ class OPDxSurfaceTest(unittest.TestCase):
 
         self.assertEqual(channel_1.info['Width_value'], 47.81942809668896)
         self.assertEqual(channel_1.info['Height_value'], 35.85522403809594)
-        self.assertEqual(channel_1.info['z_scale'], 78.592625, )
+        self.assertEqual(channel_1.info['z_scale'], 78.592625)
+
+        assert channel_1.info['unit'] == 'µm'  # see GH 281
 
         # .. mandatory keys
         self.assertEqual(channel_1.name, 'Raw')
         self.assertEqual(channel_1.dim, 2)
-        self.assertAlmostEqual(channel_1.physical_sizes[0], 35.85522403809594)
-        self.assertAlmostEqual(channel_1.physical_sizes[1], 47.81942809668896)
-        self.assertAlmostEqual(channel_1.nb_grid_pts[0], 960)
-        self.assertAlmostEqual(channel_1.nb_grid_pts[1], 1280)
+        self.assertAlmostEqual(channel_1.physical_sizes[1], 35.85522403809594)
+        self.assertAlmostEqual(channel_1.physical_sizes[0], 47.81942809668896)
+        self.assertAlmostEqual(channel_1.nb_grid_pts[1], 960)
+        self.assertAlmostEqual(channel_1.nb_grid_pts[0], 1280)
 
     def test_topography(self):
         file_path = os.path.join(DATADIR, 'opdx2.OPDx')
@@ -184,13 +191,13 @@ class OPDxSurfaceTest(unittest.TestCase):
             self.assertEqual(topography.nb_grid_pts[1], 960)
 
             # Check unit
-            self.assertEqual(topography.info['unit'], 'nm')
+            self.assertEqual(topography.info['unit'], 'µm')  # see GH 281
 
             # Check an entry in the metadata
             self.assertEqual(topography.info['SequenceNumber'], 5972)
 
             # Check a height value
-            self.assertAlmostEqual(topography.heights()[0, 0], -7731.534, places=3)
+            self.assertAlmostEqual(topography.heights()[0, 0], -7.731534, places=6)
 
     def test_read_with_check(self):
         buffer = ['V', 'C', 'A', ' ', 'D', 'A', 'T', 'A', '\x01', '\x00', '\x00', 'U', '\x07', '\x00', '\x00', '\x00']
@@ -419,5 +426,29 @@ def test_opdx_txt_consistency():
     assert (t_txt.physical_sizes[1]/t_txt.physical_sizes[0] - ratio_ref) / ratio_ref < 1e-3
     assert t_opdx.nb_grid_pts == t_txt.nb_grid_pts
 
-    # opd heights are in nm, txt in m
-    npt.assert_allclose(t_opdx.detrend().heights(), t_txt.detrend().scale(1e9).heights(), rtol=1e-3, atol=1e-3)
+    # opd file's heights are in µm, txt file's heights in m
+    assert t_opdx.info['unit'] == 'µm'
+    assert t_txt.info['unit'] == 'm'
+    npt.assert_allclose(t_opdx.detrend().heights(), t_txt.detrend().scale(1e6).heights(), rtol=1e-6, atol=1e-3)
+
+    if False:
+        import matplotlib.pyplot as plt
+        fig, ax = plt.subplots()
+        plt.colorbar(ax.imshow(t_txt.scale(1e9).heights()))
+        fig2, ax2 = plt.subplots()
+        plt.colorbar(ax2.imshow(t_opdx.heights()))
+        plt.show(block=True)
+
+def test_opdx_txt_heights_lateral_consistency():
+    t_txt = read_topography(os.path.join(DATADIR, 'opdx2.txt'))
+
+    assert t_txt.info["unit"] == "m"
+
+    # the radius of the sphere should be 250 µm
+    R = 250 * 1e-6
+
+    rhoxx, rhoyy, rhoxy = t_txt.detrend(detrend_mode="curvature").curvatures
+
+
+    assert (1/rhoxx - R) / R < 0.01
+    assert (1/rhoyy - R) / R < 0.01