+++ PyCo/ChangeLog.md
@@ -2,13 +2,40 @@
 Change log for PyCo
 ===================
 
-v0.55.0 (not yet released)
---------------------------
+v0.57.0 (15May20)
+-----------------
 
-- API: Readers now report channel info in ChannelInfo class, fixes inconsistencies in reporting channel information (#190, #192, #236)
+- MAINT: Support for new, rewritten muFFT bindings
+- ENH: Bicubic interpolation of two-dimensional topography maps
+- ENH: Fourier derivative of topography
+- BUG: Computation of plastic area is now parallelized (#303)
+- BUG: Info dictionary mutable from user 
+
+v0.56.0 (26Feb20)
+-----------------
+
+- ENH: Change orientation of some readers such that all topographies
+       look like the image in Gwyddion when plotted with
+       "pcolormesh(t.heights().T)" (#295)
+- BUG: Fixes unknown unit "um" when reading mi file (#296)
+- BUG: Fixes missing channel name for mi files (#294)
+- ENH: generate self-affine random surfaces by specifying the self-affine prefactor (#261, #278, #279)
+- BUG: now fourier synthesis can generate Linescans again (#277, #279)
+
+
+v0.55.0 (14Feb20)
+-----------------
+
+- API: Readers now report channel info in ChannelInfo class,
+       fixes inconsistencies in reporting channel information (#190, #192, #236)
 - ENH: Readers report format identifier and are self-documented (#229, #238)
+- ENH: Readers now support Gwyddion's text export format for English and German locale (#230)
 - ENH: DI reader now read acquisition date and stores it in the info dictionary
 - BUG: DI reader autodetection did not work (#258)
+- BUG: Fixes orientation for DI files (#291)
+- DOC: Added notebook showing how 2D topographies can be plotted
+- TST: Added demo notebook which shows how to plot 2D topographies
+- ENH: adhesive ideal plastic simulations with Softwall system (#260, #283)
 
 v0.54.4 (20Dec19)
 -----------------